mcp-server/requirements.txt
fastmcp>=2.0
websockets>=12.0
python-dotenv>=1.0
